Casting variables to uint8_t to remove ambiguity
The wire.h library has two functions for "requestFrom", one utilising arguments of type uint8_t and the other utilising type int arguments.  The Arduino IDE compiler will warn of an ambiguous function, as "1" is not explicitly defined as uint8_t.  This fork simply typecasts "1" to remove the ambiguity and avoid warnings at compile time.  Tested to work in Arduino 1.8.15.